Renamed GenericPointer to ObjectPointer and moved get function
