+++ syft/frameworks/torch/pointers/__init__.py
@@ -1,5 +1,6 @@
-from syft.frameworks.torch.pointers.generic_pointer import GenericPointer
+from syft.frameworks.torch.pointers.object_pointer import ObjectPointer
 from syft.frameworks.torch.pointers.callable_pointer import create_callable_pointer
+from syft.frameworks.torch.pointers.callable_pointer import CallablePointer
 from syft.frameworks.torch.pointers.pointer_tensor import PointerTensor
 
-__all__ = ["GenericPointer", "create_callable_pointer", "PointerTensor"]
+__all__ = ["ObjectPointer", "CallablePointer", "create_callable_pointer", "PointerTensor"]
